+++ unit.start.lua
@@ -2,7 +2,7 @@
 
 
 -- Initiate Globals START
-version = "v0.2"
+version = "v0.2.1"
 description = " - Minimalistic HUD from Darkwinde & Expugnator"
 
 
@@ -20,13 +20,15 @@ environment = getEnvironmentName() -- Initialize environment name
 environmentID = getEnvironmentID() -- Initialize environment ID
 
 
-ftoAtmo = 3 --export: Level of Atmospheric Fuel Tank Otimization
-ftoSpace = 3 --export: Level of Space Fuel Tank Otimization
-ftoRocket = 0 --export: Level of Rocket Fuel Tank Otimization
+coo = 4 --export: Level of Container Optimization (Mining & Inventory) - 5% mass put down reduction per level
+fto = 3 --export: Level of Fuel Tank Optimization (Mining & Inventory) - 5% mass put down reduction per level
+fthAtmo = 3 --export: Level of Atmospheric Fuel Tank Handling (Piloting) - 20% more volume put down increase per level 
+fthSpace = 3 --export: Level of Space Fuel Tank Tank Handling (Piloting) - 20% more volume put down increase per level 
+fthRocket = 0 --export: Level of Rocket Fuel Tank Tank Handling (Piloting) - 20% more volume put down increase per level
 myFuelTanks = getFuelTanks() -- initialize fuel tanks
 
 
-payloadOverhead = 10 --export: Percentage safe factor to keep enougth trust to get from planet
+payloadOverhead = 25 --export: Percentage safe factor to keep enougth trust to get from planet
 vtolPlane = false --export: Is your ship a vertical take-off and landing plane without gyroscope
 
 
@@ -93,7 +95,7 @@ end
 -- Initiate Elements END
 
 -- Initiate Timers START
-unit.setTimer('HUD', 1.5)
+unit.setTimer('HUD', 0.5)
 unit.setTimer('FUEL', 5)
 -- Initiate Timers END